feat(navigation): share the nav shell and prep the flat-nav experiment
The account menu, search, Browse/Chat tabs and footer now live in one
shell, so the flat-nav experiment only varies the Browse panel content.
Everything around it is held constant across both arms.

- Rename Nav to NavBar, move ai-first/ to navbar/, and pull NavBarFooter
  and flat-nav/ in beside it
- Remove the Create button and its create-button-nav-experiment flag
- Ship starred drag-to-reorder to everyone and drop starred-reorder
- Report nav clicks through eventUsageLogic, matching both arms on
  primary, recents and tools
- Add an onItemClicked callback to ProjectTree so a caller can report a
  click in its own terms
