+++ frontend/src/layout/panel-layout/PanelLayout.tsx
@@ -3,7 +3,6 @@ import { useActions, useMountedLogic, useValues } from 'kea'
 
 import { IconX } from '@posthog/icons'
 
-import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { IconMenu } from 'lib/lemon-ui/icons'
 import { ButtonPrimitive } from 'lib/ui/Button/ButtonPrimitives'
 import { cn } from 'lib/utils/css-classes'
@@ -11,9 +10,8 @@ import { cn } from 'lib/utils/css-classes'
 import { supportTicketCounterLogic } from 'products/conversations/frontend/supportTicketCounterLogic'
 
 import { navigation3000Logic } from '../navigation-3000/navigationLogic'
-import { Nav as AiFirstNavBar } from './ai-first/Nav'
-import { PanelLayoutPanels } from './ai-first/PanelLayoutPanels'
-import { FlatNav } from './flat-nav/FlatNav'
+import { NavBar } from './navbar/NavBar'
+import { PanelLayoutPanels } from './navbar/PanelLayoutPanels'
 import { panelLayoutLogic } from './panelLayoutLogic'
 import { PROJECT_TREE_KEY } from './ProjectTree/ProjectTree'
 import { projectTreeLogic } from './ProjectTree/projectTreeLogic'
@@ -85,7 +83,6 @@ export function PanelLayout({ className }: { className?: string }): JSX.Element 
     const { showLayoutPanel, clearActivePanelIdentifier, showLayoutNavBar } = useActions(panelLayoutLogic)
     useMountedLogic(projectTreeLogic({ key: PROJECT_TREE_KEY }))
     useMountedLogic(supportTicketCounterLogic) // Start polling for unread tickets on app load
-    const isFlatNavEnabled = useFeatureFlag('FLAT_NAV')
 
     return (
         <>
@@ -132,7 +129,7 @@ export function PanelLayout({ className }: { className?: string }): JSX.Element 
                         : {}
                 }
             >
-                {isFlatNavEnabled ? <FlatNav /> : <AiFirstNavBar />}
+                <NavBar />
             </div>
 
             {/* Mobile-only positioning anchor for panel content. Decoupled from #project-panel-layout
