+++ frontend/src/layout/panel-layout/navbar/NavBar.scss
