+++ frontend/src/layout/panel-layout/navbar/PanelLayoutPanels.test.tsx
