+++ frontend/src/layout/panel-layout/navbar/PanelLayoutPanels.tsx
