+++ frontend/src/layout/panel-layout/navbar/tabs/editToolsLogic.ts
