+++ frontend/src/layout/panel-layout/navbar/tabs/flat-nav/FlatNavBrowse.scss
@@ -1,4 +1,4 @@
-.FlatNav[data-nav-density='compact'] {
+.FlatNavBrowse[data-nav-density='compact'] {
     --button-height-base: var(--button-height-xs);
 
     .button-primitive {
@@ -6,7 +6,7 @@
     }
 }
 
-.FlatNav .show-scrollbar-on-hover::-webkit-scrollbar {
+.FlatNavBrowse .show-scrollbar-on-hover::-webkit-scrollbar {
     width: 6px;
     height: 6px;
 }
