+++ frontend/src/layout/panel-layout/navbar/tabs/flat-nav/FlatNavSection.tsx
